+++ src/index.tsx
@@ -4,7 +4,7 @@ import { render } from 'react-dom';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import configureStore from './store/configureStore';
 import {Provider} from 'react-redux';
-import App from './App';
+import App from './components/App/App';
 import './index.css';
 
 const store = configureStore();
